Made fixes and added functionality of having multiple views for each run
