+++ packages/coinstac-ui/package.json
@@ -32,10 +32,6 @@
     "react": "^15.6.2",
     "react-apollo": "^1.4.16",
     "react-bootstrap": "^0.31.5",
-    "react-boxplot": "^0.4.0",
-    "react-d3": "^0.4.0",
-    "react-d3-basic": "^1.6.11",
-    "react-d3-core": "^1.3.9",
     "react-dnd": "^2.5.4",
     "react-dnd-html5-backend": "^2.5.4",
     "react-dom": "^15.6.2",
